requirements.txt
absl-py==0.13.0
aiofiles==0.7.0
aiohttp==3.7.4.post0
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
cycler==0.10.0
Cython==0.29.23
datasets==1.2.1
dill==0.3.4
efficientnet-pytorch==0.6.3
fastapi==0.65.2
filelock==3.0.12
fsspec==2021.5.0
future==0.18.2
gdown==3.13.0
google-auth==1.31.0
google-auth-oauthlib==0.4.4
grpcio==1.38.0
h11==0.12.0
huggingface-hub==0.0.8
idna==2.10
iniconfig==1.1.1
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
lightning-bolts==0.3.4
lightning-flash==0.2.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
multidict==5.1.0
multiprocess==0.70.12.2
munch==2.5.0
nltk==3.6.2
numpy==1.20.3
oauthlib==3.1.1
opencv-python-headless==4.5.2.52
packaging==20.9
pandas==1.2.4
Pillow==8.2.0
pluggy==0.13.1
pretrainedmodels==0.7.4
protobuf==3.17.3
py==1.10.0
pyarrow==4.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pydantic==1.8.2
pyDeprecate==0.3.0
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.4
python-dateutil==2.8.1
python-multipart==0.0.5
pytorch-lightning==1.3.6
pytorch-tabnet==3.1.0
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rouge-score==0.0.4
rsa==4.7.2
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.6.3
segmentation-models-pytorch==0.1.3
sentencepiece==0.1.96
six==1.16.0
starlette==0.14.2
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
threadpoolctl==2.1.0
timm==0.3.2
tokenizers==0.10.3
toml==0.10.2
torch==1.8.1
torchmetrics==0.3.2
torchvision==0.9.1
tqdm==4.49.0
transformers==4.7.0
typing-extensions==3.10.0.0
urllib3==1.26.5
uvicorn==0.14.0
Werkzeug==2.0.1
xxhash==2.0.2
yarl==1.6.3
